refactor: grouping pages together
