+++ src/pages/SettingsPage/SettingsPage.scss
