+++ src/pages/SettingsPage/SettingsPage.tsx
@@ -1,22 +1,16 @@
-import { FormEvent, useContext, useRef, useState } from "react";
-import "./SettingsPage.scss";
-import { User, UserContext } from "../App";
+import { FormEvent, useContext, useRef, useState } from 'react';
+import './SettingsPage.scss';
+import { User, UserContext } from '../../App';
 
 const SettingsPage = () => {
   const userContext = useContext(UserContext);
   const form = useRef<HTMLFormElement>(null);
-  const [userNameState, setUserNameState] = useState<
-    "error" | "success" | "default"
-  >("default");
-  const [placeOfWorkState, setPlaceOfWorkState] = useState<
-    "error" | "success" | "default"
-  >("default");
-  const [userNameHelper, setUserNameHelper] = useState<string | undefined>(
-    undefined
+  const [userNameState, setUserNameState] = useState<'error' | 'success' | 'default'>('default');
+  const [placeOfWorkState, setPlaceOfWorkState] = useState<'error' | 'success' | 'default'>(
+    'default',
   );
-  const [placeOfWorkHelper, setPlaceOfWorkHelper] = useState<
-    string | undefined
-  >(undefined);
+  const [userNameHelper, setUserNameHelper] = useState<string | undefined>(undefined);
+  const [placeOfWorkHelper, setPlaceOfWorkHelper] = useState<string | undefined>(undefined);
 
   const handleClick = () => {
     form.current?.requestSubmit();
@@ -27,26 +21,26 @@ const SettingsPage = () => {
 
     if (form.current) {
       const formData = new FormData(form.current);
-      if (formData.get("userName")?.toString() === "") {
-        setUserNameState("error");
+      if (formData.get('userName')?.toString() === '') {
+        setUserNameState('error');
         setUserNameHelper(`Can't be empty`);
-      } else if (formData.get("placeOfWork")?.toString() === "") {
-        setPlaceOfWorkState("error");
+      } else if (formData.get('placeOfWork')?.toString() === '') {
+        setPlaceOfWorkState('error');
         setPlaceOfWorkHelper(`Can't be empty`);
       } else {
-        setPlaceOfWorkState("success");
+        setPlaceOfWorkState('success');
         setUserNameHelper(`Saved!`);
 
-        setUserNameState("success");
+        setUserNameState('success');
         setPlaceOfWorkHelper(`Saved!`);
 
         const updatedUser: User = {
-          userName: formData.get("userName")?.toString() || "",
-          placeOfWork: formData.get("placeOfWork")?.toString() || "",
-          notificiations: userContext?.user.notificiations ?? []
+          userName: formData.get('userName')?.toString() || '',
+          placeOfWork: formData.get('placeOfWork')?.toString() || '',
+          notificiations: userContext?.user.notificiations ?? [],
         };
         userContext?.updateUser(updatedUser);
-        form.current.reset()
+        form.current.reset();
       }
     }
   };
